adding new styles and logic
