+++ components/ui/toast.tsx
@@ -96,7 +96,7 @@ const ToastTitle = React.forwardRef<
 >(({ className, ...props }, ref) => (
   <ToastPrimitives.Title
     ref={ref}
-    className={cn("text-sm font-semibold [&+div]:text-xs", className)}
+    className={cn("text-base font-semibold [&+div]:text-sm", className)}
     {...props}
   />
 ));
