+++ package.json
@@ -26,7 +26,8 @@
     "react-dom": "^18",
     "tailwind-merge": "^2.5.2",
     "tailwindcss-animate": "^1.0.7",
-    "use-sound": "^4.0.3"
+    "use-sound": "^4.0.3",
+    "usehooks-ts": "^3.1.0"
   },
   "devDependencies": {
     "@types/cypress": "^1.1.3",
