+++ tailwind.config.ts
@@ -13,6 +13,16 @@ const config: Config = {
   ],
   theme: {
     extend: {
+      fontSize: {
+        xs: "0.5rem",
+        sm: "0.625rem",
+        base: "0.75rem",
+        xl: "0.875rem",
+        "2xl": "1rem",
+        "3xl": "1.5rem",
+        "4xl": "2rem",
+        "5xl": "2.5rem",
+      },
       colors: {
         background: "hsl(var(--background))",
         foreground: "hsl(var(--foreground))",
