+++ utils/types.ts
@@ -1,0 +1,1 @@
+export type FeedbackColor = "accent" | "warning" | "success" | "default";
